pyproject.toml
[project]
name = "secretflow-worker"
version = "0.1.0"
description = "SecretFlow隐私计算生产模式后端服务"
readme = "README.md"
requires-python = "==3.10.16"
dependencies = [
    "secretflow==1.12.0b0",
    "celery[redis]",
    "redis",
    "pydantic",
    "pydantic-settings",
    "structlog",
    "cryptography",
    "kombu",
    "billiard",
    "python-jose[cryptography]",
    "passlib[bcrypt]",
    "flower",
    "loguru",
    "python-dotenv",
]

[tool.uv]
prerelease = "allow"

[tool.uv.sources]
sfl = { git = "https://github.com/secretflow/sfl.git", rev = "peicheng/fix_deps" }

[dependency-groups]
dev = [
    "ruff>=0.14.10",
]
